legacy/executor-service/requirements.txt
# Executor Service
# Executes code in Docker containers

# Docker SDK for container management
docker==6.1.3

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1

# Redis for event publishing
redis==5.0.1